HW02/pyproject.toml
[tool.poetry]
name = "hw02"
version = "0.1.0"
description = ""
authors = ["Jonathan Kloefper <kloepj@rpi.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pystan = "^3.9.0"
pandas = "^2.2.1"
numpy = "^1.26.4"
seaborn = "^0.13.2"
scipy = "^1.12.0"
arviz = "^0.17.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
